requirements.txt
matplotlib==3.3.4
pandas==1.1.5
numpy==1.19.5
streamlit==0.86.0
pyserial==3.5
